pyproject.toml
[project]
name = "aablty-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "aablty",email = "acrouple@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-multipart (>=0.0.12,<1.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "pillow (>=11.0.0,<12.0.0)",
    "httpx (>=0.28.1,<1.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]

[tool.poetry]
packages = [{include = "aablty_backend", from = "src"}]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
